+++ src/interfaces/http/routes/Category/category-size.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from "express"
 import type { CategorySizeController } from "../../controllers/Category/category-size.controller"
 import { CategorySizeValidator } from "../../validators/Category/category-size.validator"
+import { AuthorizationMiddleware } from "@interfaces/http/middlewares/authorization.middleware"
 
 export class CategorySizeRoutes {
   private router: Router
@@ -16,6 +17,7 @@ export class CategorySizeRoutes {
     // POST /category-sizes - Create a new category size
     this.router.post(
       "/",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:category']),
       CategorySizeValidator.createCategorySize(),
       this.categorySizeController.createCategorySize.bind(this.categorySizeController),
     )
@@ -23,6 +25,7 @@ export class CategorySizeRoutes {
     // GET /category-sizes - Get all category sizes with pagination
     this.router.get(
       "/",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:cashier']),
       CategorySizeValidator.getCategorySizes(),
       this.categorySizeController.getAllCategorySizes.bind(this.categorySizeController),
     )
@@ -30,6 +33,7 @@ export class CategorySizeRoutes {
     // GET /category-sizes/:id - Get category size by ID
     this.router.get(
       "/:id",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:cashier']),
       CategorySizeValidator.getCategorySizeById(),
       this.categorySizeController.getCategorySizeById.bind(this.categorySizeController),
     )
@@ -37,6 +41,7 @@ export class CategorySizeRoutes {
     // GET /category-sizes/category/:categoryId - Get sizes by category ID
     this.router.get(
       "/category/:categoryId",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:cashier']),
       CategorySizeValidator.getCategorySizesByCategoryId(),
       this.categorySizeController.getCategorySizesByCategoryId.bind(this.categorySizeController),
     )
@@ -44,6 +49,7 @@ export class CategorySizeRoutes {
     // PUT /category-sizes/:id - Update category size
     this.router.put(
       "/:id",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:category']),
       CategorySizeValidator.updateCategorySize(),
       this.categorySizeController.updateCategorySize.bind(this.categorySizeController),
     )
@@ -51,6 +57,7 @@ export class CategorySizeRoutes {
     // DELETE /category-sizes/:id - Delete category size
     this.router.delete(
       "/:id",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:category']),
       CategorySizeValidator.deleteCategorySize(),
       this.categorySizeController.deleteCategorySize.bind(this.categorySizeController),
     )
