+++ src/interfaces/http/routes/Category/category.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from "express"
 import type { CategoryController } from "../../controllers/Category/category.controller"
 import { CategoryValidator } from "../../validators/Category/category.validator"
+import { AuthorizationMiddleware } from "@interfaces/http/middlewares/authorization.middleware"
 
 export class CategoryRoutes {
   private router: Router
@@ -16,6 +17,7 @@ export class CategoryRoutes {
     // POST /categories - Create a new category
     this.router.post(
       "/",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:category']),
       CategoryValidator.createCategory(),
       this.categoryController.createCategory.bind(this.categoryController),
     )
@@ -23,6 +25,7 @@ export class CategoryRoutes {
     // GET /categories - Get all categories with pagination
     this.router.get(
       "/",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:cashier']),
       CategoryValidator.getCategories(),
       this.categoryController.getAllCategories.bind(this.categoryController),
     )
@@ -30,6 +33,7 @@ export class CategoryRoutes {
     // GET /categories/:id - Get category by ID
     this.router.get(
       "/:id",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:cashier']),
       CategoryValidator.getCategoryById(),
       this.categoryController.getCategoryById.bind(this.categoryController),
     )
@@ -37,6 +41,7 @@ export class CategoryRoutes {
     // PUT /categories/:id - Update category
     this.router.put(
       "/:id",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:category']),
       CategoryValidator.updateCategory(),
       this.categoryController.updateCategory.bind(this.categoryController),
     )
@@ -44,6 +49,7 @@ export class CategoryRoutes {
     // DELETE /categories/:id - Delete category
     this.router.delete(
       "/:id",
+      AuthorizationMiddleware.requireAnyPermission(['OWNER_ACCESS', 'access:category']),
       CategoryValidator.deleteCategory(),
       this.categoryController.deleteCategory.bind(this.categoryController),
     )
